DTNN-1273 Fixed remaining chanId int types

+++ bpa/pi/inc/bplib_pi.h
@@ -199,7 +199,7 @@ BPLib_Status_t BPLib_PI_ValidateConfigs(void *TblData);
  *  \return Execution status
  *  \retval BPLIB_SUCCESS Operation was successful
  */
-BPLib_Status_t BPLib_PI_Ingress(BPLib_Instance_t *Inst, uint8_t ChanId, 
+BPLib_Status_t BPLib_PI_Ingress(BPLib_Instance_t *Inst, uint32_t ChanId, 
                                                         void *AduPtr, size_t AduSize);
 
 /**
@@ -221,7 +221,7 @@ BPLib_Status_t BPLib_PI_Ingress(BPLib_Instance_t *Inst, uint8_t ChanId,
  *  \return Execution status
  *  \retval BPLIB_SUCCESS Operation was successful
  */
-BPLib_Status_t BPLib_PI_Egress(BPLib_Instance_t *Inst, uint8_t ChanId, void *AduPtr, 
+BPLib_Status_t BPLib_PI_Egress(BPLib_Instance_t *Inst, uint32_t ChanId, void *AduPtr, 
                                     size_t *AduSize, size_t BufLen, uint32_t Timeout);
 
 #endif /* BPLIB_PI_H */
